refactor OptiMask: change column index mapper dtype to int16 and remove verbosity; add performance test for solve method

+++ optimask/_optimask.py
@@ -141,7 +141,7 @@ class OptiMask:
         """
         m, n = x.shape
         iy, ix = np.empty(m * n, dtype=np.uint32), np.empty(m * n, dtype=np.uint32)
-        cols_index_mapper = -np.ones(n, dtype=np.int32)
+        cols_index_mapper = -np.ones(n, dtype=np.int16)
         rows_with_nan = np.zeros(m, dtype=np.uint32)
         n_rows_with_nan = 0
         n_cols_with_nan = 0
